requirements.txt
APScheduler==3.6.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==2.10
importlib-metadata==4.3.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
